♻️ #294 - 프로필 사진 업로드 컴포넌트 개발
- 회원가입과 프로필 수정 페이지에서 사용
