+++ src/components/profile/SelectProfileImage.tsx
@@ -1,18 +1,14 @@
 import styled from '@emotion/styled';
-import { isAxiosError } from 'axios';
 import Image from 'next/image';
 import {
   useRef,
-  type ChangeEventHandler,
   type Dispatch,
   type MouseEventHandler,
   type SetStateAction,
 } from 'react';
-import type { ErrorResponse } from 'types/response';
-import { ImagePickerIcon } from 'assets/icons';
+import { ProfileUpload } from 'components/common';
 import { DEFAULT_PROFILE_IMAGES } from 'constants/profile';
-import { useImageUpload } from 'hooks/services';
-import { ScreenReaderOnly, SVGVerticalAlignStyle } from 'styles';
+import { SVGVerticalAlignStyle } from 'styles';
 
 interface SelectProfileImageProps {
   previewImage: string;
@@ -24,26 +20,6 @@ export const SelectProfileImage = ({
   setPreviewImage,
 }: SelectProfileImageProps) => {
   const imageRef = useRef<Array<HTMLImageElement | null>>([]);
-  const { mutate: imageUploadMutate } = useImageUpload({ path: 'users' });
-
-  const handleImageFile: ChangeEventHandler<HTMLInputElement> = (e) => {
-    const { files } = e.target;
-    if (files !== null) {
-      const imageFormData = new FormData();
-      imageFormData.append('image', files[0]);
-
-      imageUploadMutate(imageFormData, {
-        onSuccess: (imgUrl) => {
-          setPreviewImage(imgUrl);
-        },
-        onError: (error) => {
-          if (isAxiosError<ErrorResponse>(error)) {
-            console.log(error);
-          }
-        },
-      });
-    }
-  };
 
   const handleDefaultProfileImage: MouseEventHandler<HTMLButtonElement> = (
     e,
@@ -57,36 +33,26 @@ export const SelectProfileImage = ({
 
   return (
     <ImageFileContainer>
-      <ImageFileLabel htmlFor="selectImageFile">
-        <ImagePickerIcon />
-      </ImageFileLabel>
-      <ImageFileInput
-        type="file"
-        id="selectImageFile"
-        accept="image/*"
-        onChange={handleImageFile}
-      />
-      <>
-        {DEFAULT_PROFILE_IMAGES.map((image, index) => {
-          const { id, url } = image;
-          return (
-            <ImageButton
-              key={`default-images-${id}`}
-              type="button"
-              onClick={handleDefaultProfileImage}
-              isActive={url === previewImage}
-            >
-              <Image
-                ref={(element) => (imageRef.current[index] = element)}
-                src={url}
-                alt={`기본 프로필 이미지 ${id}`}
-                width={60}
-                height={60}
-              />
-            </ImageButton>
-          );
-        })}
-      </>
+      <ProfileUpload onChange={setPreviewImage} />
+      {DEFAULT_PROFILE_IMAGES.map((image, index) => {
+        const { id, url } = image;
+        return (
+          <ImageButton
+            key={`default-images-${id}`}
+            type="button"
+            onClick={handleDefaultProfileImage}
+            isActive={url === previewImage}
+          >
+            <Image
+              ref={(element) => (imageRef.current[index] = element)}
+              src={url}
+              alt={`기본 프로필 이미지 ${id}`}
+              width={60}
+              height={60}
+            />
+          </ImageButton>
+        );
+      })}
     </ImageFileContainer>
   );
 };
@@ -97,21 +63,6 @@ const ImageFileContainer = styled.div`
   align-items: center;
   width: fit-content;
   margin: 12px auto 36px;
-`;
-
-const ImageFileLabel = styled.label`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 60px;
-  border-radius: 50%;
-  background-color: ${({ theme }) => theme.colors.bg_02};
-  aspect-ratio: 1;
-  cursor: pointer;
-`;
-
-const ImageFileInput = styled.input`
-  ${ScreenReaderOnly}
 `;
 
 const ImageButton = styled.button<{ isActive: boolean }>`
